backend/requirements.txt
flask==3.0.2
pymssql==2.3.0
argon2-cffi
